[FIX][fuel_tanks_cash_control,stock_fuel_tanks] - moved models and fields that depend on cash control to fuel_tanks_cash_control - views and translations
