+++ fuel_tanks_cash_control/models/__init__.py
@@ -3,3 +3,4 @@ from . import cash_control_session
 from . import fuel_move_line
 from . import stock_move
 from . import stock_picking
+from . import stock_pump
