+++ fuel_tanks_cash_control/views/stock_pump.xml
@@ -1,0 +1,25 @@
+<?xml version='1.0' encoding='utf-8'?>
+<odoo>
+    <!-- stock.pump inherit form view -->
+    <record id="stock_pump_view_form" model="ir.ui.view">
+        <field name="name">stock.pump.inherit.view.form</field>
+        <field name="model">stock.pump</field>
+        <field name="inherit_id" ref="stock_fuel_tanks.stock_pump_view_form"/>
+        <field name="arch" type="xml">
+            <xpath expr="//group[last()]" position="after">
+                <group>
+                    <field name="pump_history_ids" readonly="1">
+                        <tree>
+                            <field name="cash_control_session_id" />
+                            <field name="modified_date" />
+                            <field name="initial_qty" />
+                            <field name="final_qty" />
+                            <field name="cubic_meters" />
+                            <field name="amount" />
+                        </tree>
+                    </field>
+                </group>
+            </xpath>
+        </field>
+    </record>
+</odoo>
