+++ stock_fuel_tanks/security/ir.model.access.csv
@@ -1,3 +1,2 @@
 id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
 access_stock_pump,access_stock_pump,model_stock_pump,base.group_user,1,1,1,1
-access_stock_pump_history,access_stock_pump_history,model_stock_pump_history,base.group_user,1,1,1,1
